refactor: move mouse handling functions to MapInner
